analyses/anomaly-detection/app/requirements.txt
firebase-admin
numpy 
google-cloud-pubsub
google-cloud-storage
sklearn